+++ projects/demo/src/app/blog/demo-post.ts
@@ -1,0 +1,8 @@
+import { BlogPostBase } from 'ngx-blogdown';
+
+export interface DemoPost extends BlogPostBase {
+  date: string;
+  tagline: string;
+  author: string;
+  cover: string | null;
+}