+++ projects/ngx-blogdown/src/public-api.ts
@@ -2,6 +2,6 @@
  * Public API Surface of ng-blog
  */
 
-export * from './lib/blog.models';
+export type { BlogPostBase, BlogPost, NgBlogConfig, SeoTags } from './lib/blog.models';
 export * from './lib/blog.config';
 export * from './lib/blog.service';